# back-end/openapi.yaml
openapi: 3.0.3
info:
  title: Accounting API
  version: 0.2.0
  description: |
    English: Phase 1 auth endpoints with JWT access/refresh tokens.
    فارسی: فاز ۱؛ اندپوینت‌های احراز هویت با توکن دسترسی و رفرش.
  x-i18n:
    description:
      en: Phase 1 auth endpoints with JWT access/refresh tokens.
      fa: فاز ۱؛ اندپوینت‌های احراز هویت با توکن دسترسی و رفرش.
servers:
  - url: http://localhost:4100/api
paths:
...
  /v1/warehouses:
    get:
      summary: List warehouses
      x-i18n:
        summary:
          en: List warehouses
          fa: فهرست انبارها
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of warehouses
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Warehouse'
                  message:
                    type: string
    post:
      summary: Create warehouse
      x-i18n:
        summary:
          en: Create warehouse
          fa: ایجاد انبار
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                code:
                  type: string
              required: [name, code]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
  /v1/warehouses/{id}:
    patch:
      summary: Update warehouse
      x-i18n:
        summary:
          en: Update warehouse
          fa: به‌روزرسانی انبار
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                code:
                  type: string
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
        '404':
          description: Not found
    delete:
      summary: Delete warehouse
      x-i18n:
        summary:
          en: Delete warehouse
          fa: حذف انبار
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
        '404':
          description: Not found
...
  /v1/reports/trial-balance:
    get:
      summary: Trial balance
      x-i18n:
        summary:
          en: Trial balance
          fa: تراز آزمایشی
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: fiscal_year_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trial balance data
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TrialBalanceItem'
                  totals:
                    type: object
                    properties:
                      debit:
                        type: number
                      credit:
                        type: number
                  message:
                    type: string
  /v1/reports/ledger:
    get:
      summary: Ledger by account
      x-i18n:
        summary:
          en: Ledger by account
          fa: دفتر کل بر اساس حساب
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: fiscal_year_id
          required: true
          schema:
            type: string
        - in: query
          name: account_id
          required: true
          schema:
            type: string
        - in: query
          name: start_date
          required: false
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Ledger items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/LedgerItem'
                  message:
                    type: string
  /v1/reports/balance-sheet:
    get:
      summary: Balance sheet
      x-i18n:
        summary:
          en: Balance sheet
          fa: ترازنامه
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: fiscal_year_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Balance sheet summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    $ref: '#/components/schemas/BalanceSheetSummary'
                  message:
                    type: string
  /v1/reports/profit-loss:
    get:
      summary: Profit & loss
      x-i18n:
        summary:
          en: Profit & loss
          fa: سود و زیان
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: fiscal_year_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Profit & loss summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    $ref: '#/components/schemas/ProfitLossSummary'
                  message:
                    type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Warehouse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
      required: [id, name, code]
...
    TrialBalanceItem:
      type: object
      properties:
        account_id:
          type: string
        code:
          type: string
        name:
          type: string
        type:
          type: string
        debit:
          type: number
        credit:
          type: number
      required: [account_id, code, name, type, debit, credit]
    LedgerItem:
      type: object
      properties:
        journal_id:
          type: string
        date:
          type: string
          format: date
        ref_no:
          type: string
          nullable: true
        item_id:
          type: string
        debit:
          type: number
        credit:
          type: number
        description:
          type: string
          nullable: true
      required: [journal_id, date, item_id, debit, credit]
    BalanceSheetSummary:
      type: object
      properties:
        assets:
          type: number
        liabilities:
          type: number
        equity:
          type: number
        retained_earnings:
          type: number
        liabilities_plus_equity:
          type: number
      required: [assets, liabilities, equity, retained_earnings, liabilities_plus_equity]
    ProfitLossSummary:
      type: object
      properties:
        revenue:
          type: number
        expense:
          type: number
        profit:
          type: number
      required: [revenue, expense, profit]